backend/requirements.txt
numpy
matplotlib
Flask
Flask-Security-Too
Flask-SQLAlchemy
Flask-Caching
flask-restful
bcrypt
email_validator
pybase64
SQLAlchemy
Jinja2
weasyprint
celery
Flask-Cors